Update CSV test to reflect change in subject refernce and constructor. Add test for Service extractor.
